fix: remove sync external store usage
BREAKING: This change switches from `useSyncExternalStore` to vanilla useEffect for subscribing/re-subscribing to an observable. This breaks the previous behavior of having synchronously emitted values form the observable being available on first render. This means that **only** the initial value provided as the second argument to useObservable will be returned from the hook on first render. This is a breaking change in behavior; previosly any value that was _synchronously_ emitted from the observable would be available as the first return value, even if no initialValue was passed to `useObservable`.