fix(useObservable)!: eagerly subscribe with delayed refcount expiry  (#115)
* fix(useObservable): minor typing fix

* fix(test): add failing test for strict mode unsubscribe issue

* fix(useObservable)!: eagerly subscribe with delayed refcount expiry

This changes how useObservable eagerly subscribes to given observables. The observable is now set up with reference counting that expires in the next tick, which simplifies how the sync value is pulled out of it and avoids excessive subscribe + unsubscribe passes. This fixes an issue in React Strict Mode, causing subscription leaks for observables created inside a useMemo(). See strictmode.test.ts for a failing test covering this issue.

BREAKING CHANGE:
This should normally not cause any problems, but given that it introduces a change in behavior wrt. timing of observable unsubscription, it can potentially cause race conditions in rare cases, especially if side-effects are performed in response to observable emissions.
The modifications required to make the test suite pass should be a good indication of the breaking nature of these changes, although the tests suites are in the majority of cases asserting internal behavior.